restructure schema validation files
