+++ packages/tc-schema-validator/validation/ad-hoc/graphql-defs.js
@@ -1,8 +1,8 @@
 const { parse } = require('graphql');
-const sdk = require('./sdk');
+const sdk = require('../sdk');
 
 try {
-	parse(sdk.getGraphqlDefs().join('\n'))).not.toThrow()
+	parse(sdk.getGraphqlDefs().join('\n'))
 } catch (error) {
 	console.error('GraphQL schema construction failed')
 	console.error(error)
