+++ packages/tc-schema-validator/validation/ad-hoc/relationship-consistency.js
@@ -1,0 +1,5 @@
+	// Check all relationships are defined at both ends "getTwinnedRelationship"
+	// Check all relationshipTypes are referenced from both ends "describe('relationship endpoints'"
+	// Check each relationship defines direction in one place only: on the relationshipType if not reflexive, otherwise on the type
+
+const sdk = require('../sdk')
