+++ packages/tc-schema-validator/validation/json-schema/index.js
@@ -1,0 +1,30 @@
+const { enumsSchema } = require('./enums');
+const { stringPatternsSchema } = require('./string-patterns');
+const { primitiveTypesSchema } = require('./primitive-types');
+
+const getJsonSchema = () => {
+	// These depend on the sdk already being initialised
+	const { typeSchema, relationshipTypeSchema } = require('./type');
+	const { typeHierarchySchema } = require('./type-hierarchy');
+
+
+	return {
+		type: 'object',
+		properties: {
+			enums: enumsSchema,
+			types: {
+				type: 'array',
+				items: typeSchema,
+			},
+			relationshipTypes: {
+				type: 'array',
+				items: relationshipTypeSchema,
+			},
+			typeHierarchy: typeHierarchySchema,
+			stringPatterns: stringPatternsSchema,
+			primitiveTypes: primitiveTypesSchema,
+		},
+	};
+}
+
+module.exports = {getJsonSchema}
