+++ packages/tc-schema-validator/validation/json-schema/string-patterns.js
