+++ packages/tc-schema-validator/validation/json-schema/type-hierarchy.js
